Revert "Code cleanup"
This reverts commit 8b21c559a49c6deb1948e678727475c0c353f1bd.
