+++ CHANGELOG.md
@@ -1,20 +1,17 @@
-## v0.5
-- Internal code cleanup
+# Changelog
 
 ## v0.4
-- Added quick character switching for a smoother gameplay experience. (also needed for some puzzles)
-- Fixed a bug that prevented upgraded vehicles from being saved.
+- Added quick character switching for smoother gameplay.
+- Fixed a bug preventing upgraded vehicles from being saved.
 
 ## v0.3
-- Introduced favorites feature to the selection screen for quick access to characters.
-- Introduced saving of upgraded vehicles.  
-  *(Hold the OK button to save or favorite)*
-- Added UI indicators for Minfigs and Vehicles to the emulator screen next to the Token's name
+- Introduced favorites feature for quick access to characters.
+- Enabled saving of upgraded vehicle configurations.
 
 ## v0.2
-- Added vehicle placement functionality on the toypad.
+- Implemented vehicle placement functionality.
 - Introduced a new selection screen for easier navigation.
 
 ## v0.1
-- Implemented character placement on the toypad.
-- Implemented the ToyPad connection between the app and the game.
+- Implemented character placement.
+- Connected character placement to the toypad system.
