+++ application.fam
@@ -9,7 +9,7 @@ App(
         "gui",
     ],
     order=60,
-    fap_version="0.5",
+    fap_version="0.4",
     fap_icon="ldtoypad.png",  # 10x10 1-bit PNG
     fap_description="USB Toy Pad emulator for Lego Dimensions",
     fap_author="Seger",
