+++ manifest.yml
@@ -2,24 +2,12 @@ sourcecode:
   type: git
   location:
     origin: https://github.com/SegerEnd/Flipper-Zero-LD-Toypad-Emulator.git
-    commit_sha: FILL_IN_COMMIT_SHA
+    commit_sha: 8b47aac29755db2f06bc102ebe422277e44d6458
 short_description: Toy Pad emulator for Lego Dimensions via USB
 description: |
-  Use your Flipper as a ToyPad!
+  Use your Flipper as a ToyPad!!
   This app emulates the behavior of the original USB ToyPad.
   Allowing you to place characters, vehicles all directly from your Flipper!
-
-  **Favorite minifigs** and **save vehicles (including upgrades)** for quick access later.  
-  Hold the **OK** button on a character or vehicle in the emulation screen to do this.
-
-  Enable _Quick Minifig Switching_ to quickly move a minifig to a new position by placing a new one.
-  _Note: While this setting is on, it disables placing multiple copies of the same minifig on the ToyPad._
-
-  **Compatibility:**
-  Works with **PS5, PS4, PS3, and Wii U**.
-  _Xbox is not available yet._
-
-  By SegerEnd
 changelog: "@CHANGELOG.md"
 screenshots:
   - screenshots/toypad.png
